all: Make controller client an interface
This moves `controller.Client` into an interface to allow using a
different client depending on the server version being communicated with.
