+++ cli/provider.go
@@ -22,7 +22,7 @@ Commands:
 `)
 }
 
-func runProvider(args *docopt.Args, client *controller.Client) error {
+func runProvider(args *docopt.Args, client controller.Client) error {
 	if args.Bool["add"] {
 		return runProviderAdd(args, client)
 	}
@@ -45,7 +45,7 @@ func runProvider(args *docopt.Args, client *controller.Client) error {
 	return nil
 }
 
-func runProviderAdd(args *docopt.Args, client *controller.Client) error {
+func runProviderAdd(args *docopt.Args, client controller.Client) error {
 	name := args.String["<name>"]
 	url := args.String["<url>"]
