+++ cli/redis.go
@@ -37,7 +37,7 @@ Examples:
 `)
 }
 
-func runRedis(args *docopt.Args, client *controller.Client) error {
+func runRedis(args *docopt.Args, client controller.Client) error {
 	config, err := getAppRedisRunConfig(client)
 	if err != nil {
 		return err
@@ -53,7 +53,7 @@ func runRedis(args *docopt.Args, client *controller.Client) error {
 	return nil
 }
 
-func getAppRedisRunConfig(client *controller.Client) (*runConfig, error) {
+func getAppRedisRunConfig(client controller.Client) (*runConfig, error) {
 	appRelease, err := client.GetAppRelease(mustApp())
 	if err != nil {
 		return nil, fmt.Errorf("error getting app release: %s", err)
@@ -61,7 +61,7 @@ func getAppRedisRunConfig(client *controller.Client) (*runConfig, error) {
 	return getRedisRunConfig(client, mustApp(), appRelease)
 }
 
-func getRedisRunConfig(client *controller.Client, app string, appRelease *ct.Release) (*runConfig, error) {
+func getRedisRunConfig(client controller.Client, app string, appRelease *ct.Release) (*runConfig, error) {
 	redisApp := appRelease.Env["FLYNN_REDIS"]
 	if redisApp == "" {
 		return nil, fmt.Errorf("No redis server found. Provision one with `flynn resource add redis`")
@@ -84,7 +84,7 @@ func getRedisRunConfig(client *controller.Client, app string, appRelease *ct.Rel
 	return config, nil
 }
 
-func runRedisCLI(args *docopt.Args, client *controller.Client, config *runConfig) error {
+func runRedisCLI(args *docopt.Args, client controller.Client, config *runConfig) error {
 	config.Entrypoint = []string{"redis-cli"}
 	config.Env["PAGER"] = "less"
 	config.Env["LESS"] = "--ignore-case --LONG-PROMPT --SILENT --tabs=4 --quit-if-one-screen --no-init --quit-at-eof"
@@ -92,7 +92,7 @@ func runRedisCLI(args *docopt.Args, client *controller.Client, config *runConfig
 	return runJob(client, *config)
 }
 
-func runRedisDump(args *docopt.Args, client *controller.Client, config *runConfig) error {
+func runRedisDump(args *docopt.Args, client controller.Client, config *runConfig) error {
 	config.Stdout = os.Stdout
 	if filename := args.String["--file"]; filename != "" {
 		f, err := os.Create(filename)
@@ -118,7 +118,7 @@ func runRedisDump(args *docopt.Args, client *controller.Client, config *runConfi
 	return runJob(client, *config)
 }
 
-func runRedisRestore(args *docopt.Args, client *controller.Client, config *runConfig) error {
+func runRedisRestore(args *docopt.Args, client controller.Client, config *runConfig) error {
 	config.Stdin = os.Stdin
 	var size int64
 	if filename := args.String["--file"]; filename != "" {
