+++ controller/examples/examples.go
@@ -13,6 +13,7 @@ import (
 	"time"
 
 	cc "github.com/flynn/flynn/controller/client"
+	"github.com/flynn/flynn/controller/client/v1"
 	ct "github.com/flynn/flynn/controller/types"
 	"github.com/flynn/flynn/discoverd/client"
 	"github.com/flynn/flynn/host/types"
@@ -25,7 +26,7 @@ import (
 
 type generator struct {
 	conf        *config
-	client      *cc.Client
+	client      cc.Client
 	recorder    *httprecorder.Recorder
 	resourceIds map[string]string
 }
@@ -277,7 +278,7 @@ func (e *generator) deleteRoute() {
 func (e *generator) deleteApp() {
 	// call Delete rather than DeleteApp as the latter uses the app stream
 	// to watch app_deletion events.
-	e.client.Delete(fmt.Sprintf("/apps/%s", e.resourceIds["app"]), nil)
+	e.client.(*v1controller.Client).Delete(fmt.Sprintf("/apps/%s", e.resourceIds["app"]), nil)
 }
 
 func (e *generator) createArtifact() {
@@ -497,7 +498,7 @@ func (e *generator) listDeployments() {
 }
 
 func (e *generator) eventsList() {
-	events, err := e.client.ListEvents(cc.ListEventsOptions{
+	events, err := e.client.ListEvents(ct.ListEventsOptions{
 		Count: 10,
 	})
 	if err != nil {
@@ -511,7 +512,7 @@ func (e *generator) eventsList() {
 
 func (e *generator) eventsStream() {
 	events := make(chan *ct.Event)
-	e.client.StreamEvents(cc.StreamEventsOptions{
+	e.client.StreamEvents(ct.StreamEventsOptions{
 		Past:  true,
 		Count: 10,
 	}, events)
