+++ controller/worker/release_cleanup/handler.go
@@ -14,11 +14,11 @@ import (
 
 type context struct {
 	db     *postgres.DB
-	client *controller.Client
+	client controller.Client
 	logger log15.Logger
 }
 
-func JobHandler(db *postgres.DB, client *controller.Client, logger log15.Logger) func(*que.Job) error {
+func JobHandler(db *postgres.DB, client controller.Client, logger log15.Logger) func(*que.Job) error {
 	return (&context{db, client, logger}).HandleReleaseCleanup
 }
