+++ pkg/backup/backup.go
@@ -9,7 +9,7 @@ import (
 	ct "github.com/flynn/flynn/controller/types"
 )
 
-func Run(client *controller.Client, out io.Writer, progress ProgressBar) error {
+func Run(client controller.Client, out io.Writer, progress ProgressBar) error {
 	tw := NewTarWriter("flynn-backup-"+time.Now().UTC().Format("2006-01-02_150405"), out, progress)
 	defer tw.Close()
 
@@ -57,7 +57,7 @@ func Run(client *controller.Client, out io.Writer, progress ProgressBar) error {
 	return nil
 }
 
-func getApps(client *controller.Client) (map[string]*ct.ExpandedFormation, error) {
+func getApps(client controller.Client) (map[string]*ct.ExpandedFormation, error) {
 	// app -> required for backup
 	apps := map[string]bool{
 		"postgres":   true,
